+++ x/uibc/quota/keeper/quota_test.go
@@ -11,19 +11,19 @@ func TestGetQuotas(t *testing.T) {
 	s := initKeeperTestSuite(t)
 	ctx, k := s.ctx, s.app.UIbcQuotaKeeper
 
-	quotas, err := k.GetAllQuotas(ctx)
+	quotas, err := k.GetAllOutflows(ctx)
 	assert.NilError(t, err)
 	assert.Equal(t, len(quotas), 0)
 
 	setQuotas := sdk.DecCoins{sdk.NewInt64DecCoin("test_uumee", 10000)}
 
-	k.SetDenomQuotas(ctx, setQuotas)
-	quotas, err = k.GetAllQuotas(ctx)
+	k.SetOutflows(ctx, setQuotas)
+	quotas, err = k.GetAllOutflows(ctx)
 	assert.NilError(t, err)
 	assert.DeepEqual(t, setQuotas, quotas)
 
 	// get the quota of denom
-	quota, err := k.GetQuota(ctx, setQuotas[0].Denom)
+	quota, err := k.GetOutflows(ctx, setQuotas[0].Denom)
 	assert.NilError(t, err)
 	assert.Equal(t, quota.Denom, setQuotas[0].Denom)
 }
@@ -40,16 +40,16 @@ func TestResetQuota(t *testing.T) {
 	ctx, k := s.ctx, s.app.UIbcQuotaKeeper
 
 	umeeQuota := sdk.NewInt64DecCoin("uumee", 1000)
-	k.SetDenomQuota(ctx, umeeQuota)
-	q, err := k.GetQuota(ctx, umeeQuota.Denom)
+	k.SetDenomOutflow(ctx, umeeQuota)
+	q, err := k.GetOutflows(ctx, umeeQuota.Denom)
 	assert.NilError(t, err)
 	assert.DeepEqual(t, q, umeeQuota)
 
 	// reset the quota
-	k.ResetQuota(ctx)
+	k.ResetAllQuotas(ctx)
 
 	// check the quota after reset
-	q, err = k.GetQuota(ctx, umeeQuota.Denom)
+	q, err = k.GetOutflows(ctx, umeeQuota.Denom)
 	assert.NilError(t, err)
 	assert.DeepEqual(t, q.Amount, sdk.NewDec(0))
 }