+++ CHANGELOG.md
@@ -46,12 +46,16 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+## Features 
+
+- [1867](https://github.com/umee-network/umee/pull/1867) Add the registered token info by `base_denom` to cli and grpc web.
+
+
 ## [v4.1.0](https://github.com/umee-network/umee/releases/tag/v4.1.0) - 2023-02-15
 
 ### Features
 
 - [1808](https://github.com/umee-network/umee/pull/1808) Blacklisted tokens automatically cleared from token registry if they have not yet been supplied.
-- [1867](https://github.com/umee-network/umee/pull/1867) Add the registered token info by `base_denom` to cli and grpc web.
 
 ### Fixes