+++ CHANGELOG.md
@@ -48,8 +48,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Features 
 
-- [1867](https://github.com/umee-network/umee/pull/1867) Add the registered token info by `base_denom` to cli and grpc web.
-
+- [1867](https://github.com/umee-network/umee/pull/1867) Allow `/denom` option on registered tokens query to get only a single token by `base_denom`.
 
 ## [v4.1.0](https://github.com/umee-network/umee/releases/tag/v4.1.0) - 2023-02-15